requirements.txt
Flask
bcrypt
pyjwt[crypto]
Faker
pytest
pytest-playwright
mysql-connector-python
python-dotenv
marshmallow